Fix bug for not showing proper Command Box and Main Window UI
